Add some utility functions to ERFA if missing
The utility functions `ae2hd`, `hd2ae`, and `hd2pa` are not included in
ERFA 0.6.3.  The code for these three functions has been transcribed
into Julia source files and they are included into the ERFA module at
runtime if the loaded ERFA module does not already defined these
functions.  This is a complete hack and should be removed once ERFA is
updated to a more modern version that already includes these functions.