Support weakdeps and bump version to 0.5.0
`Pkg` in Julia 1.9 supports "weakdeps" and "package extensions" as an
improved way to do what `Requires.jl` does.  `RadioInterferometry` now
uses these new features of Julia 1.9 while still using `Requires.jl` on
earlier Julia versions.  The extended functionality is available when
both `RadioInterferometry` and `Symbolics` are loaded from the same
environment.  See the Julia 1.9 `Pkg` docs for more details.

This also adds a `[compat]` entries for deps (and weakdeps) and actually
fixes a bug in the added functionality, which is effectively a breaking
change since the broken functionality is no longer present.  For this
reason the version has been bumped to v0.5.0.